LR5/src/main/java/ru/ssau/tk/fruktovi/sandboxoop/operations/RightSteppingDifferentialOperator.java
package ru.ssau.tk.fruktovi.sandboxoop.operations;

import ru.ssau.tk.fruktovi.sandboxoop.functions.MathFunction;

public class RightSteppingDifferentialOperator extends SteppingDifferentialOperator{
    public RightSteppingDifferentialOperator(double step) {
        super(step);
    }

    @Override
    public MathFunction derive(MathFunction function) {
        return new MathFunction() {
            @Override
            public double apply(double x) {
                return (function.apply(x + step) - function.apply(x)) / step;
            }
        };
    }

    @Override
    public double apply(double x) {
        throw new UnsupportedOperationException();
    }
}
